+++ asf_search/__init__.py
@@ -1,9 +1,9 @@
 from importlib.metadata import PackageNotFoundError, version
 
+from .exceptions import *
 from .constants import *
 from .health import *
 from .search import *
-from .baseline import *
 
 try:
     __version__ = version(__name__)