+++ asf_search/constants/DATASET/__init__.py
@@ -1,3 +1,0 @@
-"""Datasets to be used in search and related functions"""
-
-from .DATASET import *