+++ asf_search/search/search.py
@@ -1,20 +1,25 @@
-from typing import Union, Iterable
+from typing import Union, Iterable, Tuple
 import requests
+from requests.exceptions import HTTPError
 import datetime
-import json
-import asf_search
+import math
+from .results import ASFSearchResults
+from ..exceptions import ASFSearch4xxError, ASFSearch5xxError, ASFServerError
+from ..constants import INTERNAL
+from importlib.metadata import PackageNotFoundError, version
 
 
 def search(
-        absoluteOrbit: Iterable[Union[int, range]] = None,
-        asfFrame: Iterable[Union[int, range]] = None,
+        absoluteOrbit: Iterable[Union[int, Tuple[int, int]]] = None,
+        asfFrame: Iterable[Union[int, Tuple[int, int]]] = None,
         beamMode: Iterable[str] = None,
         collectionName: Iterable[str] = None,
         end: Union[datetime.datetime, str] = None,
         flightDirection: Iterable[str] = None,
-        frame: Iterable[Union[int, range]] = None,
+        frame: Iterable[Union[int, Tuple[int, int]]] = None,
         granule_list: Iterable[str] = None,
         groupID: Iterable[str] = None,
+        insarStackId: str = None,
         instrument: Iterable[str] = None,
         intersectsWith: str = None,
         lookDirection: Iterable[str] = None,
@@ -23,14 +28,13 @@ def search(
         processingDate: Union[datetime.datetime, str] = None,
         processingLevel: Iterable[str] = None,
         product_list: Iterable[str] = None,
-        relativeOrbit: Iterable[Union[int, range]] = None,
+        relativeOrbit: Iterable[Union[int, Tuple[int, int]]] = None,
         start: Union[datetime.datetime, str] = None,
         maxResults: int = None,
-        host: str = asf_search.INTERNAL.HOST,
-        output: str = 'geojson',
+        host: str = INTERNAL.HOST,
         cmr_token: str = None,
         cmr_provider: str = None
-) -> dict:
+) -> ASFSearchResults:
     """
     Performs a generic search using the ASF SearchAPI
 
@@ -43,10 +47,10 @@ def search(
     :param frame: ESA-referenced frames are offered to give users a universal framing convention. Each ESA frame has a corresponding ASF frame assigned. See also: asfframe
     :param granule_list: List of specific granules. Search results may include several products per granule name.
     :param groupID: Identifier used to find products considered to be of the same scene but having different granule names
+    :param insarStackId: Identifier used to find products of the same InSAR stack
     :param instrument: The instrument used to acquire the data. See also: platform
     :param intersectsWith: Search by polygon, linestring, or point defined in 2D Well-Known Text (WKT)
     :param lookDirection: Left or right look direction during data acquisition
-    :param maxResults: The maximum number of results to be returned by the search
     :param platform: Remote sensing platform that acquired the data. Platforms that work together, such as Sentinel-1A/1B and ERS-1/2 have multi-platform aliases available. See also: instrument
     :param polarization: A property of SAR electromagnetic waves that can be used to extract meaningful information about surface properties of the earth.
     :param processingDate: Used to find data that has been processed at ASF since a given time and date. Supports timestamps as well as natural language such as "3 weeks ago"
@@ -54,12 +58,12 @@ def search(
     :param product_list: List of specific products. Guaranteed to be at most one product per product name.
     :param relativeOrbit: Path or track of satellite during data acquisition. For UAVSAR it is the Line ID.
     :param start: Start date of data acquisition. Supports timestamps as well as natural language such as "3 weeks ago"
+    :param maxResults: The maximum number of results to be returned by the search
     :param host: SearchAPI host, defaults to Production SearchAPI. This option is intended for dev/test purposes.
-    :param output: SearchAPI output format, can be used to alter what metadata is returned and the structure of the results.
-    :param cmr_token: EDL Auth Token for authenticated searches, see https://urs.earthdata.nasa.gov/user_tokens
+    :param cmr_token: EDL authentication token for authenticated searches, see https://urs.earthdata.nasa.gov/user_tokens
     :param cmr_provider: Custom provider name to constrain CMR results to, for more info on how this is used, see https://cmr.earthdata.nasa.gov/search/site/docs/search/api.html#c-provider
 
-    :return: Dictionary of search results
+    :return: ASFSearchResults(dict) of search results
     """
 
     kwargs = locals()
@@ -91,35 +95,61 @@ def search(
         if key in data:
             data[key] = ','.join(data[key])
 
-    headers = {'User-Agent': f'{asf_search.__name__}.{asf_search.__version__}'}
+    data['output'] = 'geojson'
 
-    response = requests.post(f'https://{host}{asf_search.INTERNAL.SEARCH_PATH}', data=data, headers=headers)
+    try:
+        pkg_version = version(__name__)
+    except PackageNotFoundError:
+        pkg_version = '0.0.0'
+    headers = {'User-Agent': f'{__name__}.{pkg_version}'}
 
-    if data['output'] == 'count':
-        return {'count': int(response.text)}
-    return json.loads(response.text)
+    response = requests.post(f'https://{host}{INTERNAL.SEARCH_PATH}', data=data, headers=headers)
+
+    try:
+        response.raise_for_status()
+    except HTTPError:
+        if 400 <= response.status_code <= 499:
+            raise ASFSearch4xxError(f'HTTP {response.status_code}: {response.json()["error"]["report"]}')
+        if 500 <= response.status_code <= 599:
+            raise ASFSearch5xxError(f'HTTP {response.status_code}: {response.json()["error"]["report"]}')
+        raise ASFServerError
+
+    return ASFSearchResults(response.json())
 
 
-def flatten_list(items: Iterable[Union[int, range]]) -> str:
+def flatten_list(items: Iterable[Union[float, Tuple[float, float]]]) -> str:
     """
-    Converts a list of ints and/or ranges to a string of comma-separated ints and/or ranges.
-    Example: [1,2,3,range(4,10)] -> '1,2,3,4-10'
+    Converts a list of numbers and/or min/max tuples to a string of comma-separated numbers and/or ranges.
+    Example: [1,2,3,(10,20)] -> '1,2,3,10-20'
 
-    :param items: The list of ints and/or ranges to flatten
+    :param items: The list of numbers and/or min/max tuples to flatten
 
-    :return: String containing comma-separated representation of input, ranges converted to 'start-stop' format
+    :return: String containing comma-separated representation of input, min/max tuples converted to 'min-max' format
 
-    :raises ValueError: if input list contains non-int and non-range values, or if a range in the input list has a Step
-    != 1, or if a range in the input list is descending
+    :raises ValueError: if input list contains tuples with fewer or more than 2 values, or if a min/max tuple in the input list is descending
+    :raises TypeError: if input list contains non-numeric values
     """
 
     for item in items:
-        if isinstance(item, range):
-            if item.step != 1:
-                raise ValueError(f'Step must be 1 when using ranges to search: {item}')
-            if item.start > item.stop:
-                raise ValueError(f'Start must be less than Stop when using ranges to search: {item}')
-        elif not isinstance(item, int):
-            raise ValueError(f'Expected int or range, got {type(item)}')
+        if isinstance(item, tuple):
+            if len(item) < 2:
+                raise ValueError(f'Not enough values in min/max tuple: {item}')
+            if len(item) > 2:
+                raise ValueError(f'Too many values in min/max tuple: {item}')
+            if not isinstance(item[0], (int, float, complex)) and not isinstance(item[0], bool):
+                raise TypeError(f'Expected numeric min in tuple, got {type(item[0])}: {item}')
+            if not isinstance(item[1], (int, float, complex)) and not isinstance(item[1], bool):
+                raise TypeError(f'Expected numeric max in tuple, got {type(item[1])}: {item}')
+            if math.isinf(item[0]) or math.isnan(item[0]):
+                raise ValueError(f'Expected finite numeric min in min/max tuple, got {item[0]}: {item}')
+            if math.isinf(item[1]) or math.isnan(item[1]):
+                raise ValueError(f'Expected finite numeric max in min/max tuple, got {item[1]}: {item}')
+            if item[0] > item[1]:
+                raise ValueError(f'Min must be less than max when using min/max tuples to search: {item}')
+        elif isinstance(item, (int, float, complex)) and not isinstance(item, bool):
+            if math.isinf(item) or math.isnan(item):
+                raise ValueError(f'Expected finite numeric value, got {item}')
+        elif not isinstance(item, (int, float, complex)) and not isinstance(item, bool):
+            raise TypeError(f'Expected number or min/max tuple, got {type(item)}')
 
-    return ','.join([f'{item.start}-{item.stop}' if isinstance(item, range) else f'{item}' for item in items])
+    return ','.join([f'{item[0]}-{item[1]}' if isinstance(item, tuple) else f'{item}' for item in items])